requirements.txt
streamlit==1.31.0
semantic-kernel==0.9.5b1
anthropic==0.18.1
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.1
aiohttp==3.9.3
pydantic==2.6.1
pytest==8.0.0
pytest-asyncio==0.23.4